python-backend/requirements.txt
# FastAPI and server
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1

# Market data APIs
yfinance==0.2.50
aiohttp==3.11.11

# AI/ML APIs
anthropic>=0.20.0

# Technical analysis
ta==0.11.0
pandas==2.2.3
numpy==2.2.1

# Firebase/Firestore
google-cloud-firestore==2.20.0
firebase-admin==6.6.0

# NLP / Sentiment
vaderSentiment==3.3.2

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.20

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
httpx==0.28.1

# Development
black==24.10.0
flake8==7.1.1
mypy==1.14.1
